improve: move authenticate manager into a separate file
