+++ src/copilot-chat/src/copilot_agent/utils/__init__.py
@@ -1,6 +1,6 @@
 """Init file for utils module."""
 
-from .authentication import authenticate_user
+from .authentication import AuthenticationManager
 from .classify import QuestionClassifier
 from .conversation_manager import contextualize_question
 from .dcw import dcw_parser, extract_dcw_from_history, gen_dcw, parse_and_align_dcw
@@ -55,6 +55,7 @@ from .utils import (
 
 __all__ = [
     'DCW',
+    'AuthenticationManager',
     'AdditionalData',
     'BList',
     'Benchmark',
@@ -68,7 +69,6 @@ __all__ = [
     'SQLManager',
     'SystemData',
     'RestAPIClient',
-    'authenticate_user',
     'contextualize_question',
     'dcw_parser',
     'execute_openpai_query',
